feat(api)!: remove unstable evaluation endpoints
