+++ web/.structure-baseline.json
@@ -969,17 +969,6 @@
    "src/features/evals/server/evalTemplateDeletion.ts -> src/features/audit-logs/auditLog.ts",
    "src/features/evals/server/router.ts -> src/features/audit-logs/auditLog.ts",
    "src/features/evals/server/router.ts -> src/features/rbac/utils/checkProjectAccess.ts",
-   "src/features/evals/server/unstable-public-api/adapters.ts -> src/features/public-api/types/unstable-public-evals-contract.ts",
-   "src/features/evals/server/unstable-public-api/evaluation-rule-service.ts -> src/features/audit-logs/auditLog.ts",
-   "src/features/evals/server/unstable-public-api/evaluation-rule-service.ts -> src/features/public-api/server/unstable-public-api-error-contract.ts",
-   "src/features/evals/server/unstable-public-api/evaluation-rule-service.ts -> src/features/public-api/types/unstable-evaluation-rules.ts",
-   "src/features/evals/server/unstable-public-api/evaluator-service.ts -> src/features/audit-logs/auditLog.ts",
-   "src/features/evals/server/unstable-public-api/evaluator-service.ts -> src/features/public-api/server/unstable-public-api-error-contract.ts",
-   "src/features/evals/server/unstable-public-api/evaluator-service.ts -> src/features/public-api/types/unstable-evaluators.ts",
-   "src/features/evals/server/unstable-public-api/evaluator-service.ts -> src/features/public-api/types/unstable-public-evals-contract.ts",
-   "src/features/evals/server/unstable-public-api/types.ts -> src/features/public-api/types/unstable-public-evals-contract.ts",
-   "src/features/evals/server/unstable-public-api/validation.ts -> src/features/public-api/server/unstable-public-api-error-contract.ts",
-   "src/features/evals/server/unstable-public-api/validation.ts -> src/features/public-api/types/unstable-public-evals-contract.ts",
    "src/features/events/components/CategoryPresetChips.tsx -> src/features/posthog-analytics/usePostHogClientCapture.ts",
    "src/features/events/components/EventsTable.tsx -> src/features/batch-actions/components/AddObservationsToDatasetDialog/index.tsx",
    "src/features/events/components/EventsTable.tsx -> src/features/batch-actions/components/RunEvaluationDialog/index.tsx",
@@ -1198,28 +1187,6 @@
    "src/features/mcp/features/datasets/tools/upsertDataset.ts -> src/features/public-api/types/datasets.ts",
    "src/features/mcp/features/datasets/tools/upsertDatasetItem.ts -> src/features/datasets/server/publicDatasetService.ts",
    "src/features/mcp/features/datasets/tools/upsertDatasetItem.ts -> src/features/public-api/types/datasets.ts",
-   "src/features/mcp/features/evals/schema.ts -> src/features/public-api/types/unstable-public-evals-contract.ts",
-   "src/features/mcp/features/evals/tools/createEvaluationRule.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/createEvaluationRule.ts -> src/features/public-api/types/unstable-evaluation-rules.ts",
-   "src/features/mcp/features/evals/tools/createEvaluationRule.ts -> src/features/public-api/types/unstable-public-evals-contract.ts",
-   "src/features/mcp/features/evals/tools/deleteEvaluationRule.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/deleteEvaluationRule.ts -> src/features/public-api/types/unstable-evaluation-rules.ts",
-   "src/features/mcp/features/evals/tools/deleteEvaluator.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/deleteEvaluator.ts -> src/features/public-api/types/unstable-evaluators.ts",
-   "src/features/mcp/features/evals/tools/getEvaluationRule.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/getEvaluationRule.ts -> src/features/public-api/types/unstable-evaluation-rules.ts",
-   "src/features/mcp/features/evals/tools/getEvaluator.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/getEvaluator.ts -> src/features/public-api/types/unstable-evaluators.ts",
-   "src/features/mcp/features/evals/tools/listEvaluationRules.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/listEvaluationRules.ts -> src/features/public-api/types/unstable-evaluation-rules.ts",
-   "src/features/mcp/features/evals/tools/listEvaluators.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/listEvaluators.ts -> src/features/public-api/types/unstable-evaluators.ts",
-   "src/features/mcp/features/evals/tools/updateEvaluationRule.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/updateEvaluationRule.ts -> src/features/public-api/types/unstable-evaluation-rules.ts",
-   "src/features/mcp/features/evals/tools/updateEvaluationRule.ts -> src/features/public-api/types/unstable-public-evals-contract.ts",
-   "src/features/mcp/features/evals/tools/upsertEvaluator.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/upsertEvaluator.ts -> src/features/public-api/types/unstable-evaluators.ts",
-   "src/features/mcp/features/evals/tools/upsertEvaluator.ts -> src/features/public-api/types/unstable-public-evals-contract.ts",
    "src/features/mcp/features/experiments/schema.ts -> src/features/public-api/types/experiments.ts",
    "src/features/mcp/features/experiments/tools/listExperimentItems.ts -> src/features/experiments/server/public/service.ts",
    "src/features/mcp/features/experiments/tools/listExperimentItems.ts -> src/features/public-api/types/experiments.ts",
@@ -1830,15 +1797,6 @@
    "src/features/mcp/features/datasets/tools/listDatasets.ts -> src/features/datasets/server/publicDatasetService.ts",
    "src/features/mcp/features/datasets/tools/upsertDataset.ts -> src/features/datasets/server/publicDatasetService.ts",
    "src/features/mcp/features/datasets/tools/upsertDatasetItem.ts -> src/features/datasets/server/publicDatasetService.ts",
-   "src/features/mcp/features/evals/tools/createEvaluationRule.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/deleteEvaluationRule.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/deleteEvaluator.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/getEvaluationRule.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/getEvaluator.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/listEvaluationRules.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/listEvaluators.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/updateEvaluationRule.ts -> src/features/evals/server/unstable-public-api/index.ts",
-   "src/features/mcp/features/evals/tools/upsertEvaluator.ts -> src/features/evals/server/unstable-public-api/index.ts",
    "src/features/mcp/features/experiments/tools/listExperimentItems.ts -> src/features/experiments/server/public/service.ts",
    "src/features/mcp/features/experiments/tools/listExperiments.ts -> src/features/experiments/server/public/service.ts",
    "src/features/mcp/features/feedback/tools/submitFeedback.ts -> src/features/feedback/server/FeedbackService.ts",
