--- web/src/features/evals/server/unstable-public-api/index.ts
@@ -1,13 +1,0 @@
-export {
-  createPublicEvaluationRule,
-  deletePublicEvaluationRule,
-  getPublicEvaluationRule,
-  listPublicEvaluationRules,
-  updatePublicEvaluationRule,
-} from "./evaluation-rule-service";
-export {
-  createPublicEvaluator,
-  deletePublicEvaluator,
-  getPublicEvaluator,
-  listPublicEvaluators,
-} from "./evaluator-service";
