+++ web/src/features/public-api/server/evaluation/evaluationAdapters.ts
@@ -64,7 +64,15 @@ const INTERNAL_MAPPING_COLUMN_TO_PUBLIC_SOURCE: Record<
   experiment_item_metadata: "experiment_item_metadata",
 };
 
-function toStoredMappingList(mappings: PromptVariableMappingInputType[]) {
+export function toPublicEvaluatorType(type: EvalTemplateType) {
+  return type === EvalTemplateType.CODE
+    ? PUBLIC_EVALUATOR_TYPE_CODE
+    : PUBLIC_EVALUATOR_TYPE_LLM_AS_JUDGE;
+}
+
+export function toStoredMappingList(
+  mappings: PromptVariableMappingInputType[],
+) {
   return observationVariableMappingList.parse(
     mappings.map((mapping) => ({
       templateVariable: mapping.variable,
@@ -75,7 +83,9 @@ function toStoredMappingList(mappings: PromptVariableMappingInputType[]) {
   );
 }
 
-function toApiReadMappings(mappings: unknown): PromptVariableMappingReadType[] {
+export function toApiReadMappings(
+  mappings: unknown,
+): PromptVariableMappingReadType[] {
   const parsed = observationVariableMappingList.safeParse(mappings);
   if (!parsed.success) {
     logger.error("Failed to parse public evaluation rule mappings", {
