+++ web/src/features/public-api/server/index.ts
@@ -162,39 +162,3 @@ export {
   PostDashboardPlacementBody,
   PostUnstableDashboardBody,
 } from "../types/unstable-dashboards";
-export type {
-  PatchUnstableEvaluationRuleBodyType,
-  PostUnstableEvaluationRuleBodyType,
-} from "../types/unstable-evaluation-rules";
-export type { PostUnstableEvaluatorBodyParsedType } from "../types/unstable-evaluators";
-export type {
-  LegacyPromptVariableMappingType,
-  PromptVariableMappingInputType,
-  PromptVariableMappingReadType,
-  PublicCodeEvaluatorSourceCodeLanguageType,
-  PublicEvaluationRuleEvaluatorReferenceType,
-  PublicEvaluationRuleEvaluatorType,
-  PublicEvaluationRuleFilterType,
-  PublicEvaluationRuleReadFilterType,
-  PublicEvaluationRuleLegacyTargetType,
-  PublicEvaluationRuleReadTargetType,
-  PublicEvaluationRuleStatusType,
-  PublicEvaluationRuleTargetType,
-  PublicEvaluatorModelConfigType,
-  PublicEvaluatorOutputDefinitionType,
-  PublicEvaluatorTypeType,
-} from "../types/unstable-public-evals-contract";
-export {
-  EXPERIMENT_EVALUATION_RULE_FILTER_COLUMNS,
-  ExperimentPromptVariableMappingSource,
-  LegacyPromptVariableMapping,
-  OBSERVATION_EVALUATION_RULE_FILTER_COLUMNS,
-  ObservationPromptVariableMappingInput,
-  ObservationPromptVariableMappingSource,
-  PromptVariableMappingRead,
-  PUBLIC_EVALUATOR_TYPE_CODE,
-  PUBLIC_EVALUATOR_TYPE_LLM_AS_JUDGE,
-  PublicEvaluationRuleFilter,
-  PublicEvaluationRuleReadFilter,
-  PublicEvaluatorType,
-} from "../types/unstable-public-evals-contract";
